refactor(gateway): extract session-config + telegram-topics mixins from run.py (shard s2)
